+++ src/predict_modality/starter_kit/starter_kit_r/script.R
@@ -1,15 +1,27 @@
+# Dependencies:
+#   python: anndata
+#   r: anndata, lmds, FNN
+#
+# R starter kit for the NeurIPS 2021 Single-Cell Competition. Parts
+# with `TODO` are supposed to be changed by you.
+#
+# More documentation:
+#
+# https://viash.io/docs/creating_components/r/
+
 cat("Loading dependencies\n")
 library(anndata, warn.conflicts = FALSE, quietly = TRUE)
 library(Matrix, warn.conflicts = FALSE, quietly = TRUE)
 library(lmds, warn.conflicts = FALSE, quietly = TRUE)
+library(FNN, warn.conflicts = FALSE, quietly = TRUE)
 
 ## VIASH START
 # Anything within this block will be removed by viash
 # and will be replaced with the parameters as specified in
 # your config.vsh.yaml.
 par <- list(
-  input_mod1 = "sample_data/pbmc_1k_protein_v3.mod1.h5ad",
-  input_mod2 = "sample_data/pbmc_1k_protein_v3.mod2.h5ad",
+  input_mod1 = "sample_data/test_resource.mod1.h5ad",
+  input_mod2 = "sample_data/test_resource.mod2.h5ad",
   distance_method = "spearman",
   output = "output.h5ad",
   n_pcs = 4L,
@@ -17,13 +29,17 @@ par <- list(
 )
 ## VIASH END
 
-method_id = "mymethod" # fill in the name of your method here
+method_id <- "mymethod" # fill in the name of your method here
 
 cat("Reading h5ad files\n")
 ad1 <- read_h5ad(par$input_mod1)
 ad2 <- read_h5ad(par$input_mod2)
 
+# TODO: implement own method
+
 cat("Performing dimensionality reduction on the mod1 values\n")
+# LMDS is more efficient than regular MDS because
+# it does not compure a square distance matrix.
 dr <- lmds(
   ad1$X,
   ndim = par$n_pcs,
@@ -31,15 +47,16 @@ dr <- lmds(
 )
 
 # split up the train vs. test dimensionality reduction
-dr_train <- dr[ad1$obs$group == "train",]
-dr_test <- dr[ad1$obs$group == "test",]
+dr_train <- dr[ad1$obs$group == "train", ]
+dr_test <- dr[ad1$obs$group == "test", ]
 responses_train <- ad2$X
 
 cat("Run KNN regression.\n")
-# For every column in mod2, predict mod2 for the test cells using the K nearest mod2 neighbors
+# For every column in mod2, predict mod2 for the test cells
+# using the K nearest mod2 train neighbors
 preds <- apply(responses_train, 2, function(yi) {
   FNN::knn.reg(
-    train = dr_train, 
+    train = dr_train,
     test = dr_test,
     y = yi,
     k = par$n_neighbors
@@ -47,8 +64,9 @@ preds <- apply(responses_train, 2, function(yi) {
 })
 
 cat("Creating output matrix\n")
+# store prediction as a sparse matrix
 prediction <- Matrix::Matrix(
-  preds, 
+  preds,
   sparse = TRUE,
   dimnames = list(rownames(dr_test), colnames(ad2))
 )
@@ -63,5 +81,4 @@ out <- anndata::AnnData(
 )
 
 cat("Writing predictions to file\n")
-out$write_h5ad(par$output, compression = "gzip")
-
+zzz <- out$write_h5ad(par$output, compression = "gzip")